backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
django-debug-toolbar
gunicorn
Pillow
